pyproject.toml
[project]
name = "yourlocaljareds"
version = "0.1.0"
description = "Local RAG implementation with Qdrant, Mistral, and Open WebUI"
requires-python = ">=3.10"
dependencies = [
    # Embeddings
    "sentence-transformers>=3.0.0",

    # LLM (Ollama API)
    "requests>=2.31.0",

    # Vector DB
    "qdrant-client>=1.9.0",

    # Document parsing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.23",
    "unstructured>=0.13.0",

    # API server
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "langchain-text-splitters>=0.2.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
ylj = "start:main"
ylj-ingest = "ylj.ingest:main"
ylj-serve = "ylj.server:main"

[tool.setuptools.packages.find]
include = ["ylj*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
